moved RotatingLogger from util to logger crate
