+++ rpc/src/lib.rs
@@ -34,6 +34,7 @@ extern crate ethstore;
 extern crate ethsync;
 extern crate ethash;
 extern crate ethcore_light as light;
+extern crate ethcore_logger;
 extern crate transient_hashmap;
 extern crate jsonrpc_ipc_server as ipc;
 extern crate ethcore_ipc;
