+++ rpc/src/v1/types/transaction_request.rs
@@ -18,7 +18,7 @@
 
 use v1::types::{Bytes, H160, U256, TransactionCondition};
 use v1::helpers;
-use util::log::Colour;
+use util::Colour;
 
 use std::fmt;
