+++ util/Cargo.toml
@@ -11,11 +11,9 @@ build = "build.rs"
 log = "0.3"
 env_logger = "0.3"
 rustc-serialize = "0.3"
-arrayvec = "0.3"
 rand = "0.3.12"
 time = "0.1.34"
 rocksdb = { git = "https://github.com/ethcore/rust-rocksdb" }
-lazy_static = "0.2"
 eth-secp256k1 = { git = "https://github.com/ethcore/rust-secp256k1" }
 rust-crypto = "0.2.34"
 elastic-array = { git = "https://github.com/ethcore/elastic-array" }
@@ -32,7 +30,7 @@ ethcore-bigint = { path = "bigint" }
 parking_lot = "0.3"
 using_queue = { path = "using_queue" }
 table = { path = "table" }
-ansi_term = "0.7"
+ansi_term = "0.9"
 tiny-keccak= "1.0"
 ethcore-bloom-journal = { path = "bloom" }
 regex = "0.1"
